requirements.txt
tkinter
qrcode
pillow
opencv-python
